Re-hauled Dash Arrows, created Globals, and tested parallax
Dash Arrows now give the player a custom velocity for a certain amount of time (or until the player hits another Dash Arrow). Previously, it was force-based.
There is now a class for global variables that can be seen and set by any script in any scene.
I added a few more tricks to the gravity scene, including the new Dash Arrows and the parallax background that Muhender coded.